pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Patrick Arminio", email = "patrick.arminio@gmail.com"},
]
dependencies = [
    "strawberry-graphql>=0.206.0",
    "fastapi>=0.103.1",
    "uvicorn>=0.23.2",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

